Simple Calculator
Hello! As I've journeyed through the basics of C++ my next goal is to be well versed in the fundamentals of HTML and JavaScript. This is my second beginner project in my journey through JavaScript which is a simple calculator. In the far future I plan to learn about data structures and algorithms, once I am confident in comprehending and finding a natural flow with my coding fundamentals then I will proceed to learn the complexity of data structures and algorithms. Thank you for stopping by.